src/main.py
import torch
from torch.utils.data import DataLoader
import yaml
from model import FactorVAE
from loss import FactorVAELoss
from discriminator import Discriminator

# Load configuration
with open('../config/config.yaml', 'r') as file:
    config = yaml.safe_load(file)

# Hyperparameters
input_channels = config['input_channels']
latent_dim = config['latent_dim']
output_channels = config['output_channels']
gamma = config['gamma']
learning_rate = config['learning_rate']
batch_size = config['batch_size']
epochs = config['epochs']

# Initialize model, loss, and discriminator
model = FactorVAE(input_channels, latent_dim, output_channels)
loss_fn = FactorVAELoss(gamma)
discriminator = Discriminator(latent_dim)

# Optimizers
optimizer_vae = torch.optim.Adam(model.parameters(), lr=learning_rate)
optimizer_disc = torch.optim.Adam(discriminator.parameters(), lr=learning_rate)

# DataLoader (placeholder, replace with actual dataset)
train_loader = DataLoader([], batch_size=batch_size, shuffle=True)

# Training loop
for epoch in range(epochs):
    for batch in train_loader:
        # Forward pass
        x = batch  # Placeholder for input data
        recon_x, mu, logvar, z = model(x)

        # Compute loss
        loss, recon_loss, kl_div, tc_loss = loss_fn.compute_loss(recon_x, x, mu, logvar, z, discriminator)

        # Backward pass and optimization for VAE
        optimizer_vae.zero_grad()
        loss.backward()
        optimizer_vae.step()

        # Update discriminator
        optimizer_disc.zero_grad()
        tc_loss.backward()
        optimizer_disc.step()

    print(f'Epoch [{epoch+1}/{epochs}], Loss: {loss.item():.4f}, Recon Loss: {recon_loss.item():.4f}, KL Div: {kl_div.item():.4f}, TC Loss: {tc_loss.item():.4f}')

# Save model checkpoint
torch.save(model.state_dict(), '../results/model_checkpoint.pth')

# Note: This script is a template and requires actual dataset and configuration values to run properly.
# Ensure the dataset is loaded correctly and the config.yaml file is set up with the necessary parameters.
